Further consolidate [CodeAsset] specific classes by moving them to lib/src/code_assets
This moves more of [CodeAsset] specific things into
lib/src/code_assets/*.

But the public API surface of the package isn't changed by this PR.
