+++ pkgs/native_assets_cli/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.9.1-wip
 
 - Update pubspec.yaml of examples to use 0.9.0 of package:native_assets_cli
+- Consolidate [CodeAsset] specific things into `lib/src/code_assets/*`
 
 ## 0.9.0
