+++ pkgs/native_assets_cli/lib/src/code_assets/architecture.dart
