+++ pkgs/native_assets_cli/lib/src/code_assets/c_compiler_config.dart
@@ -4,8 +4,8 @@
 
 import 'package:collection/collection.dart';
 
-import 'json_utils.dart';
-import 'utils/map.dart';
+import '../json_utils.dart';
+import '../utils/map.dart';
 
 /// The configuration for a C toolchain.
 final class CCompilerConfig {
