+++ pkgs/native_assets_cli/lib/src/code_assets/ios_sdk.dart
