+++ pkgs/native_assets_cli/lib/src/code_assets/link_mode.dart
@@ -2,8 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'code_assets/code_asset.dart';
-import 'config.dart';
+import '../config.dart';
+
+import 'code_asset.dart';
 
 /// The link mode for a [CodeAsset].
 ///
