+++ pkgs/native_assets_cli/lib/src/target.dart
@@ -5,8 +5,8 @@
 import 'dart:ffi' show Abi;
 import 'dart:io';
 
-import '../architecture.dart';
-import '../os.dart';
+import 'code_assets/architecture.dart';
+import 'os.dart';
 
 final class Target implements Comparable<Target> {
   final Abi abi;
